Move EscapePath to util
