+++ pkg/util/filepathutil/escape.go
@@ -1,4 +1,4 @@
-package configsource
+package filepathutil
 
 import (
 	"regexp"
